[SW-320] Unblock H2OConf test
